fix: showing wrong amount on dashboard

+++ backend/src/main/java/com/dashboard/backend/dto/PaidAmountDTO.java
@@ -1,6 +1,4 @@
 package com.dashboard.backend.dto;
 
-import java.math.BigDecimal;
-
-public record PaidAmountDTO(BigDecimal paidAmount) {
+public record PaidAmountDTO(long paidAmount) {
 }
