+++ backend/src/main/java/com/dashboard/backend/dto/PendingAmountDTO.java
@@ -1,6 +1,4 @@
 package com.dashboard.backend.dto;
 
-import java.math.BigDecimal;
-
-public record PendingAmountDTO(BigDecimal pendingAmount) {
+public record PendingAmountDTO(long pendingAmount) {
 }
